requirements.txt
# requirements.txt for CV + WebRTC project
opencv-python==4.8.1.78
numpy==1.24.3
aiohttp==3.9.1
websockets==12.0
aiortc==1.6.0
av==9.2.0
cryptography==41.0.3
pyopenssl==23.2.0
asyncio==3.4.3
python-json-logger==2.0.7
python-dotenv